JD5JB-7 districtWageHandler and it's implementations refactored
