+++ app/src/test/java/com/infoshareacademy/jbusters/data/DistrWagesHandlerTest.java
@@ -7,6 +7,7 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Properties;
 
 
 public class DistrWagesHandlerTest {
@@ -16,14 +17,13 @@ public class DistrWagesHandlerTest {
 
         DistrWagesHandler distr = new DistrWagesHandler(null);
 
-        Assert.assertEquals(new HashMap<String,Integer>(), distr.getDistrictWages());
+        Assert.assertEquals(new Properties(), distr.getProperties());
     }
     @Test
     public void wrongFileTypeInputTest(){
 
         DistrWagesHandler distr = new DistrWagesHandler("pom.xml");
-        Assert.assertEquals(new HashMap<String,Integer>(), distr.getDistrictWages());
-
+        Assert.assertEquals(new Properties(), distr.getProperties());
     }
